requirements.txt
streamlit
langchain-groq
python-dotenv
speechrecognition
streamlit-mic-recorder
pyperclip